server/requirements.txt
flask
flask-cors
pymongo
python-dotenv
langchain
langchain-community
openai
anthropic
python-jose
